+++ packages/docs/site/src/routes/components/[slug].svelte
@@ -218,9 +218,10 @@
 	}
 
 	.col2 {
-		grid-column: 2 / span 1;
 		border: 1px solid lightgrey;
 		box-shadow: 1px 1px 4px 1px lightgrey;
+		grid-column: 2 / span 1;
+		overflow-y: auto;
 	}
 	.col2 .svgwrapper {
 		height: 100%;